feat(web): init profile component and finalize scaffold of login comp
