+++ packages/nextjs/components/Header/components/Web3AuthConnectButton.tsx
@@ -3,7 +3,7 @@ import toast from "react-hot-toast";
 import { useWeb3AuthContext } from "~~/context/Web3AuthContext";
 
 export const Web3AuthConnectButton = () => {
-  const { connectWeb3Auth, disconnectWeb3Auth, initWeb3Auth, isConnected } = useWeb3AuthContext();
+  const { connectWeb3Auth, initWeb3Auth } = useWeb3AuthContext();
 
   const openConnectModal = async () => {
     try {
@@ -12,16 +12,6 @@ export const Web3AuthConnectButton = () => {
     } catch (error) {
       console.error(error);
       toast.error("Error while trying to connect. Please try again");
-    }
-  };
-
-  const logOut = async () => {
-    try {
-      await disconnectWeb3Auth();
-      toast.success("Successfully logged out");
-    } catch (error) {
-      console.error(error);
-      toast.error("Error while trying to disconnect. Please try again");
     }
   };
 
@@ -38,11 +28,7 @@ export const Web3AuthConnectButton = () => {
     initAuth();
   }, [initWeb3Auth]);
 
-  return isConnected ? (
-    <button className="btn btn-primary btn-sm" onClick={logOut} type="button">
-      Log out
-    </button>
-  ) : (
+  return (
     <button className="btn btn-primary btn-sm" onClick={openConnectModal} type="button">
       Log in
     </button>
